deploying web - 3

+++ db/migrate/20221021065544_create_reviews.rb
@@ -3,7 +3,7 @@ class CreateReviews < ActiveRecord::Migration[7.0]
     create_table :reviews, id: false, primary_key: [:product_id, :user_id] do |t|
       t.uuid :product_id, null: false
       t.uuid :user_id, null: false
-      t.smallint :user_score_rating
+      t.integer :user_score_rating
       t.boolean :is_favored, null: false, default: false
 
       t.timestamps
